+++ src/TRegx/CleanRegex/Pattern.php
@@ -21,9 +21,9 @@ class Pattern
     /** @var InternalPattern */
     private $pattern;
 
-    public function __construct(string $pattern, string $flags = '')
+    private function __construct(InternalPattern $pattern)
     {
-        $this->pattern = InternalPattern::automatic($pattern, $flags);
+        $this->pattern = $pattern;
     }
 
     public function test(string $subject): bool
@@ -97,7 +97,19 @@ class Pattern
 
     public static function of(string $pattern, string $flags = ''): Pattern
     {
-        return new Pattern($pattern, $flags);
+        return new Pattern(InternalPattern::standard($pattern, $flags));
+    }
+
+    /**
+     * @param string $delimitedPattern
+     * @return Pattern
+     * @deprecated Please use method \TRegx\CleanRegex\Pattern::of. Method Pattern::pcre() is only present, in case
+     * if there's an automatic delimiters' bug, that would make "Pattern" error-prone.
+     * @see \TRegx\CleanRegex\Pattern::of
+     */
+    public static function pcre(string $delimitedPattern): Pattern
+    {
+        return new Pattern(InternalPattern::pcre($delimitedPattern));
     }
 
     public static function prepare(array $input, string $flags = ''): Pattern