+++ test/Integration/TRegx/CleanRegex/CompositePattern/anyMatches/CompositePatternTest.php
@@ -4,6 +4,7 @@ namespace Test\Integration\TRegx\CleanRegex\CompositePattern\anyMatches;
 use PHPUnit\Framework\TestCase;
 use TRegx\CleanRegex\CompositePattern;
 use TRegx\CleanRegex\Internal\CompositePatternMapper;
+use TRegx\CleanRegex\Pattern;
 
 class CompositePatternTest extends TestCase
 {
@@ -13,7 +14,7 @@ class CompositePatternTest extends TestCase
     public function shouldMatch()
     {
         // given
-        $pattern = new CompositePattern((new CompositePatternMapper(['/https?/i', 'fail', '/failed/']))->createPatterns());
+        $pattern = new CompositePattern((new CompositePatternMapper($this->patterns()))->createPatterns());
 
         // when
         $match = $pattern->anyMatches('http');
@@ -28,12 +29,21 @@ class CompositePatternTest extends TestCase
     public function shouldNotMatch()
     {
         // given
-        $pattern = new CompositePattern((new CompositePatternMapper(['/https?$/i', 'fail', '/failed/']))->createPatterns());
+        $pattern = new CompositePattern((new CompositePatternMapper($this->patterns()))->createPatterns());
 
         // when
-        $match = $pattern->anyMatches('httpz');
+        $match = $pattern->anyMatches('Foo');
 
         // then
         $this->assertFalse($match);
+    }
+
+    private function patterns(): array
+    {
+        return [
+            Pattern::pcre('/https?/i'),
+            Pattern::of('fail'),
+            Pattern::pcre('/failed/')
+        ];
     }
 }